+++ modules-services-metrics/service-metrics/src/main/java/org/openprovenance/prov/service/metrics/MetricsCalculator.java
@@ -12,6 +12,7 @@ import org.openprovenance.prov.rules.SimpleMetrics;
 import org.openprovenance.prov.rules.TrafficLight;
 import org.openprovenance.prov.rules.TrafficLightResult;
 import org.openprovenance.prov.rules.counters.EntityActivityDerivationCounter;
+import org.openprovenance.prov.scala.summary.TypePropagator;
 import org.openprovenance.prov.scala.typemap.IncrementalProcessor;
 import org.openprovenance.prov.service.validation.ValidationObjectMaker;
 import org.openprovenance.prov.validation.Config;
@@ -50,7 +51,7 @@ public class MetricsCalculator extends Rules {
 
 
     @Override
-    public Object getMetrics(Document document, ProvFactory pFactory) {
+    public Object computeMetrics(Document document, ProvFactory pFactory) {
 
 
         Tuple2<Map<String, Integer>, Map<List<Tuple2<Integer, Integer>>, Integer>> features=ip.process(document);
@@ -64,38 +65,58 @@ public class MetricsCalculator extends Rules {
         Object validationReport = getValidationReportOrError(document, pFactory);
         Object traffic = getTrafficLightOrError(metrics);
 
+        String id=null;
 
         try{
-
-
-            String id=querier.insertMetrics("document",
+            id=querier.insertMetricsRecord("document",
                     null,
-                    om.writeValueAsString(features._1),
+                    TypePropagator.om().writeValueAsString(features._2),
                     om.writeValueAsString(metrics),
                     om.writeValueAsString(validationReport),
                     om.writeValueAsString(traffic));
             System.out.println("=========== ID: " + id);
+
             persistTypeMap();
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
         }
+        MetricsRecord metricsRecord = new MetricsRecord(id, "document", metrics, features._1, validationReport, traffic);
+
+        return metricsRecord;
+
+        /*
+        String finalId = id;
         return new HashMap<String, Object>() {{
             put("artifact", "document");
+            put("id", finalId);
             if (metrics!=null) put("counts", metrics);
             if (features!=null && features._2!=null) put("features", features._2);
             if (validationReport!=null) put("validity", validationReport);
+            if (traffic!=null) put("traffic", traffic);
         }};
+
+         */
     }
 
     private Object getTrafficLightOrError(Object metrics) {
         try {
             Map<String, Object> m1=(Map<String, Object>) metrics;
             Map<String, Object> m2=(Map<String, Object>)m1.get(PATTERN_METRICS);
+
             EntityActivityDerivationCounter count=(EntityActivityDerivationCounter)m2.get(COUNT_DERIVATIONS_AND_GENERATIONS_AND_USAGES);
             List<TrafficLightResult> trafficLight= TrafficLight.getTrafficLight(count);
-            return trafficLight;
+            TrafficLightResult patternTrafficLightResult=new TrafficLightResult("Pattern", trafficLight);
+
+            List<TrafficLightResult> typeTrafficLightSubResults=List.of(new TrafficLightResult("Type1", 66, TrafficLight.TrafficLightColor.ORANGE));
+            TrafficLightResult typeTrafficLightResult=new TrafficLightResult("Type",typeTrafficLightSubResults);
+
+
+            List<TrafficLightResult> trafficLightSubResults = List.of(patternTrafficLightResult, typeTrafficLightResult);
+            TrafficLightResult overallTrafficLightResult=new TrafficLightResult("Overall",trafficLightSubResults);
+
+            return overallTrafficLightResult;
         } catch (Throwable e) {
-            HashMap<String,String> trafficError = new HashMap();
+            HashMap<String,String> trafficError = new HashMap<>();
             ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
             e.printStackTrace(new PrintStream(outputStream));
             trafficError.put("error", outputStream.toString());
@@ -105,9 +126,9 @@ public class MetricsCalculator extends Rules {
 
     private Object getMetricsOrError(Document document, ProvFactory pFactory) {
         try {
-            return super.getMetrics(document, pFactory);
+            return super.computeMetrics(document, pFactory);
         } catch (Throwable e) {
-            HashMap<String,String> metricsError = new HashMap();
+            HashMap<String,String> metricsError = new HashMap<>();
             ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
             e.printStackTrace(new PrintStream(outputStream));
             metricsError.put("error", outputStream.toString());
@@ -123,7 +144,7 @@ public class MetricsCalculator extends Rules {
             validationReport.nonStrictCycle = null;
             return validationReport;
         } catch (Throwable e) {
-            HashMap<String,String> validationError = new HashMap();
+            HashMap<String,String> validationError = new HashMap<>();
             ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
             e.printStackTrace(new PrintStream(outputStream));
             validationError.put("error", outputStream.toString());